Refactor benchmark cli.
